typecast array indexing to int from char to prevent compiler warning
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@985 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/drivers/lcd.c
@@ -819,7 +819,7 @@ static void scroll_thread(void)
                 s->line[i] = s->line[i+1];
 
             if ( s->offset < s->textlen ) {
-                s->line[s->space - 1] = s->text[s->offset];
+                s->line[(int)s->space - 1] = s->text[(int)s->offset];
                 s->offset++;
             }
             else {
